senate: support voting on upgrade proposals
Page similar to the document polls one, except for contract upgrade
votes. Links to the contract's code on Etherscan.
