+++ src/lib/routeNames.js
@@ -19,6 +19,8 @@ export const ROUTE_NAMES = {
   SIGIL_GENERATOR: Symbol('SIGIL_GENERATOR'),
   CREATE_GALAXY: Symbol('CREATE_GALAXY'),
   SENATE: Symbol('SENATE'),
+  DOCUMENTS: Symbol('DOCUMENTS'),
+  UPGRADES: Symbol('UPGRADES'),
   ISSUE_CHILD: Symbol('ISSUE_CHILD'),
   PARTY_SET_POOL_SIZE: Symbol('PARTY_SET_POOL_SIZE'),
   INVITE_COHORT: Symbol('INVITE_COHORT'),
